refactor(UI,output): Move UI code to output package
- Avoid circular dependency for testing
